Categories CRUD implemented except delete, missing to change and verify habits
