+++ backend/habits/code/database/database.js
@@ -141,7 +141,9 @@ async function createHabitsTable(db) {
       id INTEGER PRIMARY KEY AUTOINCREMENT,
       name TEXT NOT NULL, 
       start_time TEXT, 
-      duration INTEGER)`,
+      duration INTEGER,
+      category_id INTEGER NOT NULL,
+      FOREIGN KEY (category_id) REFERENCES categories(id))`,
           (err) => {
             if (err) {
               console.error('Error creating table: ', err.message);
