+++ frontend/src/routes/categories/[id]/habits/[id]/+page.svelte
@@ -27,5 +27,3 @@
         </div>
     </li> 
 </ul>
-
-
