+++ frontend/src/routes/categories/[id]/habits/[id]/edit/+page.svelte
