+++ frontend/src/routes/home/+page.svelte
@@ -8,7 +8,7 @@
 <div class="h-screen flex justify-center items-center">
 	<div class="flex flex-col justify-center items-center gap-4 mb-10">
 		<div class="flex flex-row gap-7">
-			<OctopusButton buttonName='Habits' buttonIcon='habitOctopus.png' buttonLink="/habits"/>
+			<OctopusButton buttonName='Habits' buttonIcon='habitOctopus.png' buttonLink="/categories"/>
 			<OctopusButton buttonName='Moods' buttonIcon='moodOctopus.png' buttonLink='/moods'/>
 		</div>
 		<div class="bg-white w-full rounded-xl shadow-2xl px-5 py-4 text-center justify-center items-center" style="box-shadow: rgba(0, 0, 0, 0.15) 2.4px 2.4px 3.2px">
